+++ CHANGELOG.md
@@ -19,8 +19,15 @@ These changes are available on the `master` branch, but have not yet been releas
   ([#1940](https://github.com/Pycord-Development/pycord/pull/1940))
 - Added support for text-related features in `StageChannel`.
   ([#1936](https://github.com/Pycord-Development/pycord/pull/1936))
+- Added support for one-time event listeners in `@client.listen()`.
+  ([#1957](https://github.com/Pycord-Development/pycord/pull/1957))
 - Added `current_page` argument to Paginator.update()
-  ([#1983](https://github.com/Pycord-Development/pycord/pull/1983))
+  ([#1983](https://github.com/Pycord-Development/pycord/pull/1983)
+
+### Removed
+
+- Removed `@client.once()` in favour of `@client.listen(once=True)`
+  ([#1957](https://github.com/Pycord-Development/pycord/pull/1957))
 
 ## [2.4.1] - 2023-03-20